% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDemo.R
\name{runDemo}
\alias{runDemo}
\title{Application Demo}
\usage{
runDemo()
}
\arguments{
\item{...}{Nothing}
}
\value{
Launch the shiny application to give you the ability to play with the numbers
}
\description{
Application Demo
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
