% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{estimate_area}
\alias{estimate_area}
\title{Finding area}
\usage{
estimate_area(B = 5000, seed = 10)
}
\arguments{
\item{B}{number of points}

\item{seed}{for random number}
}
\value{
value of points in the area
}
\description{
Finding area
}
