% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{estimate_area}
\alias{estimate_area}
\title{Finding area using a Monte-Carlo approach}
\usage{
estimate_area(B = 5000, seed = 10)
}
\arguments{
\item{B}{number of points}

\item{seed}{for random number}
}
\value{
Return the estimated area and the coordinates of the points
}
\description{
Finding area using a Monte-Carlo approach
}
\examples{
estimate_area(50)

}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
