% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot_area}
\alias{plot_area}
\title{Plotting area}
\usage{
plot_area(x)
}
\arguments{
\item{x}{is area}
}
\value{
A plot showing the area
}
\description{
Plotting area
}
\examples{

x <- estimate_area(B = 5000, seed = 10)
plot_area(x)

}
